Refactored and organized
Kept editing all of the files and organizing it for better readability.
I removed login.php and instead have an icon in the navbar that allows
you to log in from there, thus removing the clutter of files. I
implemented a database class to ensure a session only has one database
connection up at a time and reduces the amount of duplicate code.